\name{circtest}
\alias{circtest}
\alias{circtest2}
\title{ circtest / sample data }
\description{
  Sample data for the \code{link{circumcircle}} function.

  \code{circtest2} are points sampled from a circle with some jitter
  added, i.e. they represent the most complicated case for the
  \code{link{circumcircle}} function.
  
}

\keyword{datasets}
