\name{print.summary.triSht}
\title{Print a summary of a triangulation object}
\usage{\method{print}{summary.triSht}(x, ...)}
\alias{print.summary.triSht}
\arguments{
  \item{x}{object of class \code{"summary.triSht"},
    generated by \code{\link{summary.triSht}}.}
  \item{...}{additional paramters for \code{print}}
}
\description{
  Prints some information about \code{tri.obj}
}
\value{None
}
\note{
  This function is meant as replacement for the function of same name in
  package \code{tripack}.

  The only difference is that no constraints are possible with
  \code{triSht} objects of package \code{interp}.
}
\author{
  Albrecht Gebhardt <albrecht.gebhardt@aau.at>,
  Roger Bivand <roger.bivand@nhh.no>
}
\seealso{
  \code{\link{triSht}},\code{\link{tri.mesh}},
  \code{\link{print.triSht}}, \code{\link{plot.triSht}},
  \code{\link{summary.triSht}}.
}
\keyword{spatial}
