\name{print.summary.voronoi}
\title{Print a summary of a voronoi object}

\usage{\method{print}{summary.voronoi}(x, ...)}
\alias{print.summary.voronoi}
\arguments{
  \item{x}{object of class \code{"summary.voronoi"},
    generated by \code{\link{summary.voronoi}}.}
  \item{...}{additional paramters for \code{print}}
}
\description{
  Prints some information about object \code{x}
}
\value{None
}
\note{
  This function is meant as replacement for the function of same name in
  package \code{tripack} and should be fully backward compatible.
}
\author{
  Albrecht Gebhardt <albrecht.gebhardt@aau.at>,
  Roger Bivand <roger.bivand@nhh.no>
}
\seealso{
  \code{\link{voronoi}},\code{\link{voronoi.mosaic}},
  \code{\link{print.voronoi}}, \code{\link{plot.voronoi}},
  \code{\link{summary.voronoi}}. 
}
\keyword{spatial}
