\name{print.triSht}
\title{Print a triangulation object}
\usage{\method{print}{triSht}(x,...)
}
\alias{print.triSht}
\arguments{
  \item{x}{object of class \code{"triSht"}}
  \item{...}{additional paramters for \code{print}}
}
\description{prints a adjacency list of \code{"x"}
}
\value{None
}
\author{
  Albrecht Gebhardt <albrecht.gebhardt@aau.at>,
  Roger Bivand <roger.bivand@nhh.no>
}

\seealso{
  \code{\link{triSht}},
  \code{\link{plot.triSht}},
  \code{\link{summary.triSht}}
}
\keyword{spatial}
