\name{print.voronoi}
\title{Print a voronoi object}
\usage{\method{print}{voronoi}(x,...)
}
\alias{print.voronoi}
\arguments{
  \item{x}{object of class \code{"voronoi"}}
  \item{...}{additional paramters for \code{print}}
}
\description{prints a summary of \code{"x"}
}
\value{None
}
\author{
  Albrecht Gebhardt <albrecht.gebhardt@aau.at>,
  Roger Bivand <roger.bivand@nhh.no>
}

\seealso{
  \code{\link{voronoi}},
  \code{\link{plot.voronoi}},
  \code{\link{summary.voronoi}}
}
\keyword{spatial}
