\name{tritest}
\alias{tritest}
\alias{tritest2}
\title{ tritest / sample data }
\description{
A very simply set set of points to test the tripack functions, taken
from the FORTRAN original. \code{tritest2} is a slight modification by 
adding \code{runif(,-0.1,0.1)} random numbers to the coordinates.
}
\references{
  R. J. Renka (1996). Algorithm 751: TRIPACK: a constrained
  two-dimensional Delaunay triangulation package.
  ACM Transactions on Mathematical Software.
  \bold{22}, 1-8.
}
\keyword{datasets}
